+++ plinth/modules/bind/__init__.py
@@ -22,8 +22,6 @@ Plinth module to configure BIND server
 import re
 
 from django.utils.translation import ugettext_lazy as _
-from django.core.exceptions import ValidationError
-from django.core.validators import validate_ipv46_address
 
 from plinth import actions
 from plinth import action_utils
@@ -141,13 +139,3 @@ def get_default():
         'forwarders': forwarders
     }
     return conf
-
-
-def validate(ips):
-    """Validate that ips is a list of IP addresses, separated by space."""
-    for ip_addr in ips.split():
-        try:
-            validate_ipv46_address(ip_addr)
-        except ValidationError:
-            return False
-    return True